+++ src/ai-examples/azext_ai_examples/custom.py
@@ -23,26 +23,15 @@ def check_connection_aladdin():
         print('Connection failed')
 
 
-# Replacements for core functions
-def provide_examples(help_file):
-    return replace_examples(help_file)
+def new_examples(help_file):
+    help_file.examples = replace_examples(help_file.command)
 
 
-# Provide two options for changing the examples
-
-# Replace built in examples wirh Aladdin ones
-def replace_examples(help_file):
+# Replace built in examples with Aladdin ones
+def replace_examples(command):
     # Specify az to coerce the examples to be for the exact command
-    lookup_term = "az " + help_file.command
+    lookup_term = "az " + command
     return get_generated_examples(lookup_term)
-
-
-# Append Aladdin examples to the built in ones
-def append_examples(help_file):
-    # Specify az to coerce the examples to be for the exact command
-    lookup_term = "az " + help_file.command
-    aladdin_examples = get_generated_examples(lookup_term)
-    return concat_unique_examples(help_file.examples, aladdin_examples)
 
 
 # Support functions
@@ -52,26 +41,11 @@ def get_generated_examples(cli_term):
 
     if response.status_code == 200:
         for answer in json.loads(response.content):
-            examples.append(clean_from_http_answer(answer))
+            # Ignore pruned responses
+            if answer['source'] != 'pruned':
+                examples.append(clean_from_http_answer(answer))
 
     return examples
-
-
-def concat_unique_examples(first_list, second_list):
-    for first_item in first_list:
-        for second_item in second_list:
-            if are_examples_equal(first_item, second_item):
-                second_list.remove(second_item)
-    return first_list + second_list
-
-
-def are_examples_equal(first, second):
-    return clean_string(first.short_summary) == clean_string(second.short_summary) \
-        or clean_string(first.command) == clean_string(second.command)
-
-
-def clean_string(text):
-    return text.strip()
 
 
 def clean_from_http_answer(http_answer):
@@ -104,23 +78,19 @@ def ping_aladdin_service():
 
 
 def call_aladdin_service(query):
-    client_request_id = ''
-    if telemetry_core._session.application:  # pylint: disable=protected-access
-        client_request_id = telemetry_core._session.application.data['headers']['x-ms-client-request-id']  # pylint: disable=protected-access
-
-    session_id = telemetry_core._session._get_base_properties()['Reserved.SessionId']  # pylint: disable=protected-access
+    correlation_id = telemetry_core._session.correlation_id  # pylint: disable=protected-access
     subscription_id = telemetry_core._get_azure_subscription_id()  # pylint: disable=protected-access
-    client_request_id = client_request_id  # pylint: disable=protected-access
-    installation_id = telemetry_core._get_installation_id()  # pylint: disable=protected-access
     version = str(parse_version(core_version))
 
     context = {
-        "sessionId": session_id,
-        "subscriptionId": subscription_id,
-        "clientRequestId": client_request_id,
-        "installationId": installation_id,
+        "correlationId": "",
+        "subscriptionId": "",
         "versionNumber": version
     }
+
+    # Only pull in the other values if we have consent
+    if telemetry_core.is_telemetry_enabled():
+        context.update(correlationId=correlation_id, subscriptionId=subscription_id)
 
     api_url = 'https://app.aladdin.microsoft.com/api/v1.0/examples'
     headers = {'Content-Type': 'application/json'}