+++ src/healthcareapis/HISTORY.rst
@@ -3,6 +3,16 @@
 Release History
 ===============
 
+0.2.0
++++++
+
+* Add support for managed identity and setting the export storage account
+
+0.1.3
++++++
+
+* Remove the limitation of max compatible cli core version
+
 0.1.2
 +++++